ART: Speed up stack guard page install
Only the main thread doesn't have its stack mapped in under normal
conditions. Reading each page is a lot of overhead and we should
try to avoid it.

A simple change is to respect the outcome of the preceding Unprotect
call. If the unprotect succeeded, the stack is already mapped in.

Bug: 27718174
Change-Id: I16b214567585ed2f09970f618ccdec7eed219fd3
Signed-off-by: mydongistiny <jaysonedson@gmail.com>